drivers: create virtio drivers from device tree

+++ zircon-object/src/vm/vmar.rs
@@ -221,7 +221,7 @@ impl VmAddressRegion {
         let mut flags = flags;
         // if vmo != 0
         {
-            flags |= MMUFlags::from_bits_truncate(vmo.cache_policy() as u32 as usize);
+            flags |= MMUFlags::from_bits_truncate(vmo.cache_policy() as usize);
         }
         // align = 1K? 2K? 4K? 8K? ...
         if !self.test_map(inner, offset, len, PAGE_SIZE) {
